Fixed imports after picking protocols files re-organization
