+++ pyworkflow/em/packages/relion/protocol_autopick_v2.py
@@ -28,7 +28,7 @@ import os
 from os.path import relpath
 
 import pyworkflow.protocol.params as params
-from pyworkflow.em.protocol.protocol_particles import ProtParticlePickingAuto
+from pyworkflow.em.protocol import ProtParticlePickingAuto
 from pyworkflow.em.constants import RELATION_CTF, ALIGN_NONE
 from pyworkflow.em.convert import ImageHandler
 from pyworkflow.utils.properties import Message
